유사동영상 찾기, id로 동영상 찾기 구현

+++ src/main/java/codepirate/tubelensbe/video/domain/TrendingVideo.java
@@ -36,11 +36,11 @@ public class TrendingVideo {
 
     private String channelTitle ;
 
-    private BigInteger viewCount ;
+    private Long viewCount ;
     
-    private BigInteger likeCount ;
+    private Long likeCount ;
     
-    private BigInteger commentCount ;
+    private Long commentCount ;
 
     @ElementCollection
     private List<String> tags;
